mlx5/fs: preserve the insertion order of FTE destinations
The default for mlx5 driver is to manipulate the destination list
entries in LIFO order with the most recent non-table destination added
to the beginning. Change this to FIFO ordering so that extending the
destination list of a FTE doesn't modify the ordering of existing
destinations.

Introduce an experimental module knob to toggle upstream behavior.

Signed-off-by: Kenneth Klette Jonassen <kenneth@bridgetech.tv>