mlx5/fs: improve coexistence of zero and non-zero flow tag
Driver allows creating rules with non-zero flow_tag succeeded by rules
using the same match critera (same FTE) with a zero flow_tag, but not
vice versa. Allow either ordering here so we avoid race conditions
during flow creation.

Allow a new rule to replace an existing flow_tag. This is a kludge to
avoid checking/counting which of the rules that make up an FTE had set
or did not set flow_tag. (A proper upstream patch would need to do that.)

Signed-off-by: Kenneth Klette Jonassen <kenneth@bridgetech.tv>